convert js to ts files
