+++ package.json
@@ -21,6 +21,7 @@
     "@types/express": "^4.17.17",
     "@types/jest": "^29.5.3",
     "@types/node": "^20.4.5",
+    "@types/pg": "^8.10.2",
     "@typescript-eslint/parser": "^6.2.0",
     "ts-node": "^10.9.1",
     "ts-node-dev": "^2.0.0",
