+++ src/routes/userRoutes.ts
@@ -1,0 +1,9 @@
+import { Router } from 'express';
+import * as userController from '../controllers/userController';
+
+const router = Router();
+
+router.get('/:userId/wishlist', userController.getWishlist);
+router.post('/:userId/wishlist/:restaurantId', userController.addWish);
+
+export default router;
